requirements.txt
requests>=2.28.0
websocket-client>=1.5.0
python-dotenv
